+++ core/src/discrete_wavelet_transform.c
@@ -1,16 +1,17 @@
 #include "discrete_wavelet_transform.h"
 #include <stdint.h>
+#include <stdio.h>
 #include <string.h>
 
-static inline int16_t round_int(int32_t numerator, int32_t denominator) {
+static inline int32_t round_int(int32_t numerator, int32_t denominator) {
     float rounding = ((float)numerator)/denominator + 0.5f;
-    if(rounding < 0 && rounding != (int16_t)rounding) {
-        return (int16_t)(rounding - 1);
+    if(rounding < 0 && rounding != (int32_t)rounding) {
+        return (int32_t)(rounding - 1);
     }
-    return (int16_t)(rounding);
+    return (int32_t)(rounding);
 }
 
-static void forward_DWT(int16_t* data, size_t width) {
+static void forward_DWT(int32_t* data, size_t width) {
     //An array of size W [0.....W] is transformed to highpass 'H' and lowpass 'L' values.
     //The transformation is in-place and the resulting array is formed as:
     //[L_0,L_1 ... L_(W/2), H_0, H1 ... H_(W/2)]
@@ -18,12 +19,12 @@ static void forward_DWT(int16_t* data, size_t width) {
     //The transformation is calculated as shown in (3.3.2)
 
 	//cache line for in place operation
-	int16_t cache[width];
-	memcpy(&cache, data, width * sizeof(int16_t));
+	int32_t cache[width];
+	memcpy(&cache, data, width * sizeof(int32_t));
 
 	uint32_t n = width >> 1; //number of coefficients in pass
-	int16_t* highpass = &data[n];
-	int16_t* lowpass = &data[0];
+	int32_t* highpass = &data[n];
+	int32_t* lowpass = &data[0];
 
 
     highpass[0] = cache[1] - round_int((9 * (cache[0] + cache[2]) - (cache[2] + cache[4])), 16);
@@ -41,14 +42,14 @@ static void forward_DWT(int16_t* data, size_t width) {
     lowpass[n-1] = cache[2*n-2] - round_int(-(highpass[n-2] + highpass[n-1]), 4);
 }
 
-static void backward_DWT(int16_t* data, size_t width) {
+static void backward_DWT(int32_t* data, size_t width) {
     //The original values are reconstructed as shown in (3.4.2)
 
     size_t n = width >> 1;
-    int16_t cache[width];
-	int16_t* lowpass = &cache[0];
-	int16_t* highpass = &cache[n];
-    memcpy(&cache, data, width * sizeof(int16_t));
+    int32_t cache[width];
+	int32_t* lowpass = &cache[0];
+	int32_t* highpass = &cache[n];
+    memcpy(&cache, data, width * sizeof(int32_t));
     
     data[0] = lowpass[0] + round_int(-highpass[0], 2);
 
@@ -67,7 +68,7 @@ static void backward_DWT(int16_t* data, size_t width) {
 }
 
 
-void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data_height, uint8_t transform_levels, uint8_t inverse_flag) {
+void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data_height, uint8_t transform_levels, uint8_t inverse_flag) {
     //The 2D-image is split into 1D rows and columns.
     //These 1D arrays are operated on independantly.
     //The resulting 2D lowpass area is then transformed.
@@ -79,7 +80,7 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
             uint32_t current_height = data_height >> level;
 
             //Vertical
-			int16_t temp_column[current_height];
+			int32_t temp_column[current_height];
             for(uint32_t column = 0; column < current_width; ++column) {
 
 				//Copy current column to temp_column
@@ -88,7 +89,7 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
 					temp_column[row] = data[row * data_width + column];
 				}
 
-				backward_DWT((int16_t*)&temp_column, current_height);
+				backward_DWT((int32_t*)&temp_column, current_height);
 
 				for(uint32_t row = 0;  row < current_height; ++row) {
 					data[row * data_width + column] = temp_column[row];
@@ -96,7 +97,7 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
 			}
 
             //Horizontal
-			int16_t temp_row[current_width];
+			int32_t temp_row[current_width];
 			//Iterate each row
             for(uint32_t row = 0; row < current_height; ++row) {
 
@@ -105,7 +106,7 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
 					temp_row[column] = data[row * data_width + column];
 				}
 
-                backward_DWT((int16_t*)&temp_row, current_width);
+                backward_DWT((int32_t*)&temp_row, current_width);
 
 				for(uint32_t column = 0;  column < current_width; ++column) {
 					data[row * data_width + column] = temp_row[column];
@@ -120,9 +121,8 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
 		uint32_t current_width = data_width >> level;
 		uint32_t current_height = data_height >> level;
 
-		
 		//Horizontal
-		int16_t temp_row[current_width];
+		int32_t temp_row[current_width];
 		//Iterate each row
 		for(uint32_t row = 0; row < current_height; ++row) {
 
@@ -131,7 +131,7 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
 				temp_row[column] = data[row * data_width + column];
 			}
 
-			forward_DWT((int16_t*)&temp_row, current_width);
+			forward_DWT((int32_t*)&temp_row, current_width);
 
 			for(uint32_t column = 0;  column < current_width; ++column) {
 				data[row * data_width + column] = temp_row[column];
@@ -139,7 +139,7 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
 		}
 
 		//Vertical
-		int16_t temp_column[current_height];
+		int32_t temp_column[current_height];
 		for(uint32_t column = 0; column < current_width; ++column) {
 
 			//Copy current column to temp_column
@@ -148,7 +148,7 @@ void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data
 				temp_column[row] = data[row * data_width + column];
 			}
 
-			forward_DWT((int16_t*)&temp_column, current_height);
+			forward_DWT((int32_t*)&temp_column, current_height);
 
 			for(uint32_t row = 0;  row < current_height; ++row) {
 				data[row * data_width + column] = temp_column[row];
